requirements.txt
pyqt == 4.11.4